Use closures instead of eval to construct external CF limits
This prevents an arbitrary code execution vulnerability in
__BuildLimitCheck, wherein the $quoted_value scalar would fail to escape
\'s, allowing the user-provided VALUE to escape the quoted string.
Simply replace the eval with a closure, which does not have the same
vulnerabilities.

The other use of eval, in __BuildAggregatorsCheck, does not interpolate
user-provided data, but has also been removed in the interests of
clarity and bullet-proofing.  In doing so, the interpretation of
mixed-entryaggregators has changed slightly; association is now
left-to-right, not the previous defualt of && binding tighter than ||.